some fixes

+++ routes/users.js
@@ -8,7 +8,7 @@ const { User, Article } = require('../models/index');
 const viewsScheme = require('../schemes/viewsScheme');
 const Views = mongoose.model('articles_views', viewsScheme);
 
-usersRouter.get('/', async (req, res, next) => {
+usersRouter.get('/', async (req, res) => {
   try {
     const users = await sequelize.query(
       `select users.*, COUNT(authorId) 
@@ -19,35 +19,34 @@ usersRouter.get('/', async (req, res, next) => {
         nest: true,
       });
 
-    const articlesViews = await Views.find({}, (err, result) => {
-      if (err) { return console.log(err); }
-      return result;
-    });
+    const articlesViews = await Views.find({});
 
-    const mapped = users.map(item => {
-      let count = []
-      articlesViews.find(element => {
-        if (element.authorId === item.id) {
-          count.push(element.views);
-        }
-      });
-      let reduce = count.reduce((total, amount) => total + amount);
+    const mapped = await users.map(item => {
+      if (articlesViews.length) {
+        let count = [0];
+        articlesViews.find(element => {
+          if (element.authorId === item.id) { count.push(element.views) }
+        });
+        let reduce = count.reduce((total, amount) => total + amount);
 
-      return { ...item, viewsCount: reduce };
+        return { ...item, viewsCount: reduce };
+      }
+      return { ...item }
     });
 
     res.send({ data: mapped });
-  } catch (err) { next(err) }
+
+  } catch (err) { logger.error(err, 'Error'); }
 });
 
-usersRouter.get('/:id', async (req, res, next) => {
+usersRouter.get('/:id', async (req, res) => {
   try {
     const user = await User.findOne({ where: { id: req.params.id } });
     res.send({ data: user });
-  } catch (err) { next(err) }
+  } catch (err) { logger.error(err, 'Error'); }
 });
 
-usersRouter.post('/', async (req, res, next) => {
+usersRouter.post('/', async (req, res) => {
   try {
     const user = await User.create({
       firstName: req.body.firstName,
@@ -56,10 +55,10 @@ usersRouter.post('/', async (req, res, next) => {
       password: req.body.password,
     })
     res.send({ data: user });
-  } catch (err) { next(err) }
+  } catch (err) { logger.error(err, 'Error'); }
 });
 
-usersRouter.put('/:id', async (req, res, next) => {
+usersRouter.put('/:id', async (req, res) => {
   try {
     const user = await User.update({
       firstName: req.body.firstName,
@@ -72,10 +71,10 @@ usersRouter.put('/:id', async (req, res, next) => {
       }
     })
     res.send({ data: user });
-  } catch (err) { next(err) }
+  } catch (err) { logger.error(err, 'Error'); }
 });
 
-usersRouter.get('/:id/blog', async (req, res, next) => {
+usersRouter.get('/:id/blog', async (req, res) => {
   try {
     const article = await Article.findAll({
       order: [['id', 'DESC']],
@@ -85,10 +84,7 @@ usersRouter.get('/:id/blog', async (req, res, next) => {
       nest: true,
     })
 
-    const articlesViews = await Views.find({}, (err, result) => {
-      if (err) { return console.log(err); }
-      return result;
-    });
+    const articlesViews = await Views.find({}).exec();
 
     const mapped = article.map(item => {
       const viewsElement = articlesViews.find(element => element.articleId === item.id);
@@ -96,25 +92,20 @@ usersRouter.get('/:id/blog', async (req, res, next) => {
     });
 
     res.send({ data: mapped });
-  } catch (err) { next(err) }
+  } catch (err) { logger.error(err, 'Error'); }
 });
 
-usersRouter.delete('/:id', async (req, res, next) => {
+usersRouter.delete('/:id', async (req, res) => {
   try {
     logger.info(req, `deleted ${req.params.authorId} articles`);
     const users = await User.destroy({
       where: { id: req.params.id }
     });
 
-    await Views.deleteMany({
-      authorId: req.params.id,
-    }, (err, result) => {
-      if (err) { return console.log(err); }
-      return result;
-    })
+    await Views.deleteMany({ authorId: req.params.id }).exec();
 
     res.send({ data: users });
-  } catch (err) { next(err) }
+  } catch (err) { logger.error(err, 'Error'); }
 });
 
 module.exports = usersRouter;
